units/semantics/requirements.txt
# Optional: install with `pip install -r units/semantics/requirements.txt`
# LanguageDetector unit uses lingua (Rust-backed, offline).
lingua-language-detector>=2.0.0
